move rendering of video display to a secondary thread
previously, all rendering occured on the primary thread, with the emulator
itself running on a secondary thread. this caused multithreading issues
when passing window and input events to the emulation thread. in order to
avoid each hardware device having to deal with these issues, the main
rendering of the video display was moved to a secondary thread, and the
emulator itself was moved to the main thread

the main thread now runs the emulator, handles window and input events,
and presents the offscreen framebuffers rendered to by the secondary
video thread